(temporarily?) Move substructure generation out of main codebase
