+++ utilities/make_substructure_dict/_make_chemical_substructures.py
@@ -1,7 +1,7 @@
 import os
 
 from openff.toolkit.utils import get_data_file_path
-from openff.toolkit.utils._cif_to_substructure_dict import CifSubstructures
+from _cif_to_substructure_dict import CifSubstructures
 
 if not os.path.exists("aa-variants-v1.cif"):
     import requests
@@ -20,15 +20,16 @@ if not os.path.exists("aa-variants-v1.cif"):
 cif_object = CifSubstructures()
 cif_object.from_file(
     "aa-variants-v1.cif",
-    replace_quadruple_bond_with_any=True,
-    remove_charge_bond_order_resonant=True,
+    replace_quadruple_bond_with_any=False,
+    remove_charge_bond_order_resonant=False,
 )
 
-# Automatically add known substructures that are missing from aa_variants but expected by force fields (like ACE and
-# NME caps)
+# Automatically patch known problems - better that this explodes when things are fixed
+cif_object._patch_known_problems()
 cif_object._add_common_substructures()
+cif_object._add_common_linkages()
 
 output_file = get_data_file_path("proteins/T4-protein.sdf").replace(
-    "T4-protein.sdf", "aa_residues_substructures_with_caps.json"
+    "T4-protein.sdf", "aa_residues_substructures_explicit_bond_orders_with_caps.json"
 )
 cif_object.to_json_file(output_file)
